+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/preferences/ServerSettings.java
@@ -15,6 +15,15 @@ public class ServerSettings extends JsonSettings {
   private String mappingVpsTableId = "WEBGameID";
   private String mappingVpsTableVersionId = "CUSTOM3";
   private String mappingHsFileName = "CUSTOM4";
+  private boolean stickyKeysEnabled = true;
+
+  public boolean isStickyKeysEnabled() {
+    return stickyKeysEnabled;
+  }
+
+  public void setStickyKeysEnabled(boolean stickyKeysEnabled) {
+    this.stickyKeysEnabled = stickyKeysEnabled;
+  }
 
   public boolean isKeepModificationDate() {
     return keepModificationDate;